requirements.txt
streamlit
psycopg2
python-dotenv
